extract routing module
