+++ examples/compare-tokens.rs
@@ -1,7 +1,7 @@
 use anyhow::Result;
 use scylla::cluster::NodeAddr;
 use scylla::frame::value::ValueList;
-use scylla::sharding::Token;
+use scylla::routing::Token;
 use scylla::{Session, SessionBuilder};
 use std::env;
