+++ scylla/src/cluster/metadata.rs
@@ -2,7 +2,7 @@ use crate::cluster::host_filter::HostFilter;
 use crate::cluster::resolve_contact_points;
 use crate::connection::{Connection, ConnectionConfig, NodeConnectionPool, PoolConfig, PoolSize};
 use crate::frame::response::event::Event;
-use crate::sharding::Token;
+use crate::routing::Token;
 use crate::statement::query::Query;
 use crate::transport::errors::{DbError, QueryError};
 use crate::utils::parse::{ParseErrorCause, ParseResult, ParserState};
