+++ scylla/src/cluster/node.rs
@@ -6,7 +6,7 @@ use crate::connection::Connection;
 use crate::connection::VerifiedKeyspaceName;
 use crate::connection::{NodeConnectionPool, PoolConfig};
 /// Node represents a cluster node along with it's data and connections
-use crate::sharding::{Sharder, Token};
+use crate::routing::{sharding::Sharder, Token};
 use crate::transport::errors::QueryError;
 
 use std::fmt::Display;
