+++ scylla/src/connection/connection_pool.rs
@@ -2,7 +2,10 @@
 use crate::cloud::set_ssl_config_for_scylla_cloud_host;
 
 use crate::connection::{Connection, ConnectionConfig, ErrorReceiver, VerifiedKeyspaceName};
-use crate::sharding::{Shard, ShardCount, Sharder, Token};
+use crate::routing::{
+    sharding::{Shard, ShardCount, Sharder},
+    Token,
+};
 use crate::transport::errors::QueryError;
 
 #[cfg(feature = "cloud")]
