+++ scylla/src/routing/mod.rs
@@ -1,0 +1,7 @@
+pub mod partitioner;
+pub mod sharding;
+
+#[derive(PartialEq, Eq, PartialOrd, Ord, Clone, Copy, Debug)]
+pub struct Token {
+    pub value: i64,
+}
