+++ scylla/src/routing/partitioner.rs
@@ -2,7 +2,7 @@ use bytes::Buf;
 use std::num::Wrapping;
 
 use crate::{
-    frame::value::SerializedValues, prepared_statement::TokenCalculationError, sharding::Token,
+    frame::value::SerializedValues, prepared_statement::TokenCalculationError, routing::Token,
 };
 
 #[allow(clippy::upper_case_acronyms)]
@@ -366,9 +366,7 @@ mod tests {
     use rand::Rng;
     use rand_pcg::Pcg32;
 
-    use crate::transport::partitioner::PartitionerHasher;
-
-    use super::{CDCPartitioner, Murmur3Partitioner, Partitioner};
+    use super::{CDCPartitioner, Murmur3Partitioner, Partitioner, PartitionerHasher};
 
     fn assert_correct_murmur3_hash(pk: &'static str, expected_hash: i64) {
         let hash = Murmur3Partitioner.hash_one(pk.as_bytes()).value;
