+++ scylla/src/transport/mod.rs
@@ -1,4 +1,3 @@
-pub mod partitioner;
 pub mod query_result;
 
 pub use crate::frame::{Authenticator, Compression};
