+++ scylla/tests/integration/consistency.rs
@@ -4,8 +4,8 @@ use scylla::execution::retries::FallthroughRetryPolicy;
 use scylla::execution::{ExecutionProfileBuilder, ExecutionProfileHandle};
 use scylla::load_balancing::{DefaultPolicy, LoadBalancingPolicy, RoutingInfo};
 use scylla::prepared_statement::PreparedStatement;
+use scylla::routing::Token;
 use scylla::session::Session;
-use scylla::sharding::Token;
 use scylla::test_utils::unique_keyspace_name;
 use tokio::sync::mpsc::{self, UnboundedReceiver, UnboundedSender};
